adding flag for agda memory, startup command uses the fastest time not the average, updating readme for CI, stop translating files if every language has run into an error
