+++ .github/workflows/ci.yml
@@ -1,4 +1,4 @@
-name: build
+name: Build
 
 on:
   workflow_dispatch:
